requirements.txt
streamlit
pysnow
langchain
langchain-community
chromadb
sentence-transformers
PyPDF2
requests
python-dotenv
fastapi
uvicorn
